+++ Changes.md
@@ -46,6 +46,7 @@ API
 - GraphGadget : Added `annotationsGadget()` function.
 - MetadataAlgo : Added `annotations()` variant accepting `Gaffer::Metadata::RegistrationTypes`. The default is `All` to match existing behavior and the previous `annotations()` variant is deprecated.
 - AnnotationsGadget : Added `annotationAt()` function.
+- AnnotationUI : Added `contextMenuSignal()` allowing customisations to the context menu for annotations.
 
 1.5.4.1 (relative to 1.5.4.0)
 =======